+++ docs/changelog.md
@@ -43,6 +43,9 @@ Requires libvips v8.12.1
   [#3006](https://github.com/lovell/sharp/pull/3006)
   [@christopherbradleybanks](https://github.com/christopherbradleybanks)
 
+* Ensure rotate-then-extract works with EXIF mirroring.
+  [#3024](https://github.com/lovell/sharp/issues/3024)
+
 ## v0.29 - *circle*
 
 Requires libvips v8.11.3